fallSpeed runs on doubles, created plists for all 30 levels, but (8-30 incomplete)

+++ Projectfiles/GameLayer.h
@@ -19,8 +19,8 @@
     int framecount;
     int level;
     int numberOfEnemies;
-    int minDuration;
-    int maxDuration;
+    double minDuration;
+    double maxDuration;
     int enemiesKilled;
     int deaths;
     int bar;
